refactor(thread): extract ThreadView; ThreadModal becomes Dialog wrapper
ThreadView now contains all inner state and JSX; ThreadModal is a 22-line
Dialog wrapper. Enables reusing ThreadView for the upcoming mobile thread
route page. Desktop modal behavior is pixel-identical.
